third_party/blink/web_tests/fast/js/cyclic-proto-expected.txt

This test checks that setting a cyclic value for __proto__ throws an exception and does not alter __proto__. This was reported as bug 17927.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS x.__proto__ = x; threw exception TypeError: Cyclic __proto__ value.
PASS x.__proto__ is originalProto
PASS successfullyParsed is true

TEST COMPLETE
